+++ read_json.hpp
@@ -12,5 +12,6 @@ extern int mb_geometry[GEOM_TYPES];
 
 void json_context(json_object *j);
 void parse_geometry(int t, json_object *j, drawvec &out, int op, const char *fname, int line, json_object *feature);
+std::vector<mvt_layer> parse_layers(FILE *fp, int z, unsigned x, unsigned y, int extent);
 
 serial_val stringify_value(json_object *value, const char *reading, int line, json_object *feature);